+++ angular-workspace/projects/example-client-app/src/app/customapp/customapp.component.ts
@@ -1,5 +1,6 @@
+/* eslint-disable no-alert */
 import { Component } from '@angular/core';
-import { DrawerLocation } from '@ni/nimble-angular';
+import { DrawerLocation, MenuItem } from '@ni/nimble-angular';
 
 @Component({
     selector: 'example-customapp',
@@ -7,7 +8,12 @@ import { DrawerLocation } from '@ni/nimble-angular';
     styleUrls: ['./customapp.component.scss']
 })
 export class CustomAppComponent {
-    public drawerLocation: DrawerLocation = DrawerLocation.Right;
+    public drawerLocation: DrawerLocation = DrawerLocation.right;
     public isDrawerPinned = false;
     public drawerLocations = DrawerLocation;
+
+    public onMenuButtonMenuChange(event: Event): void {
+        const menuItemText = (event.target as MenuItem).innerText;
+        alert(`${menuItemText} selected`);
+    }
 }