+++ packages/nimble-blazor/NimbleBlazor/Components/NimbleTabsToolbar.razor.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.AspNetCore.Components;
+
+namespace NimbleBlazor;
+
+public partial class NimbleTabsToolbar : ComponentBase
+{
+    [Parameter]
+    public RenderFragment? ChildContent { get; set; }
+
+    [Parameter(CaptureUnmatchedValues = true)]
+    public IDictionary<string, object>? AdditionalAttributes { get; set; }
+}