+++ packages/nimble-blazor/Tests/NimbleBlazor.Tests/Unit/Components/NimbleButtonTests.cs
@@ -1,0 +1,40 @@
+using Bunit;
+using Xunit;
+
+namespace NimbleBlazor.Tests.Unit.Components;
+
+/// <summary>
+/// Tests for <see cref="NimbleButton"/>.
+/// </summary>
+public class NimbleButtonTests
+{
+    [Fact]
+    public void NimbleButton_Render_HasButtonMarkup()
+    {
+        var context = new TestContext();
+        context.JSInterop.Mode = JSRuntimeMode.Loose;
+        var expectedMarkup = "nimble-button";
+
+        var button = context.RenderComponent<NimbleButton>();
+
+        Assert.Contains(expectedMarkup, button.Markup);
+    }
+
+    [Theory]
+    [InlineData(ButtonAppearance.Block, "block")]
+    [InlineData(ButtonAppearance.Outline, "outline")]
+    [InlineData(ButtonAppearance.Ghost, "ghost")]
+    public void ButtonAppearance_AttributeIsSet(ButtonAppearance value, string expectedAttribute)
+    {
+        var button = RenderNimbleButton(value);
+
+        Assert.Contains(expectedAttribute, button.Markup);
+    }
+
+    private IRenderedComponent<NimbleButton> RenderNimbleButton(ButtonAppearance appearance)
+    {
+        var context = new TestContext();
+        context.JSInterop.Mode = JSRuntimeMode.Loose;
+        return context.RenderComponent<NimbleButton>(p => p.Add(x => x.Appearance, appearance));
+    }
+}