+++ packages/nimble-components/src/menu-button/template.ts
@@ -1,13 +1,23 @@
-import { elements, html, ref, slotted, when } from '@microsoft/fast-element';
+import {
+    elements,
+    html,
+    ref,
+    slotted,
+    ViewTemplate,
+    when
+} from '@microsoft/fast-element';
+import type { FoundationElementTemplate } from '@microsoft/fast-foundation';
 import type { MenuButton } from '.';
+import { ToggleButton } from '../toggle-button';
+import { AnchoredRegion } from '../anchored-region';
 
 // prettier-ignore
-export const template = html<MenuButton>`
+export const template: FoundationElementTemplate<ViewTemplate<MenuButton>> = context => html<MenuButton>`
     <template
         ?open="${x => x.open}"
         @focusout="${(x, c) => x.focusoutHandler(c.event as FocusEvent)}"
     >
-        <nimble-toggle-button
+        <${context.tagFor(ToggleButton)}
             part="button"
             appearance="${x => x.appearance}"
             content-hidden="${x => x.contentHidden}"
@@ -26,11 +36,11 @@ export const template = html<MenuButton>`
             <span slot="end" part="end">
                 <slot name="end"></slot>
             </span>
-        </nimble-toggle-button>
+        </${context.tagFor(ToggleButton)}>
         ${when(
         x => x.open,
         html<MenuButton>`
-            <nimble-anchored-region
+            <${context.tagFor(AnchoredRegion)}
                 fixed-placement="true"
                 auto-update-mode="auto"
                 horizontal-inset="true"
@@ -45,7 +55,7 @@ export const template = html<MenuButton>`
                 <span part="menu">
                     <slot name="menu" ${slotted({ property: 'slottedMenus', filter: elements('[role=menu]') })}></slot>
                 </span>
-            </nimble-anchored-region>
+            </${context.tagFor(AnchoredRegion)}>
         `
     )}
     </template>