+++ packages/nimble-blazor/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @ni/nimble-blazor
 
-This log was last generated on Tue, 03 May 2022 22:31:24 GMT and should not be manually modified.
+This log was last generated on Wed, 18 May 2022 21:26:22 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 3.0.0
+
+Wed, 18 May 2022 21:26:22 GMT
+
+### Major changes
+
+- Rename NimbleBlazor.Components to NimbleBlazor ([ni/nimble@6ef01f6](https://github.com/ni/nimble/commit/6ef01f660e91240fb8381984fde864774bd6b2bb))
+
+## 2.0.4
+
+Wed, 18 May 2022 20:52:01 GMT
+
+### Patches
+
+- Update icon generation dependencies ([ni/nimble@d6fdd7c](https://github.com/ni/nimble/commit/d6fdd7ce77c264d72f885b2a734a692dc1938e71))
+
+## 2.0.0
+
+Tue, 17 May 2022 20:06:42 GMT
+
+### Major changes
+
+- Replace single Appearance enum with one per control (ButtonAppearance, TextAreaAppearance, TextFieldAppearance). Add TextFieldAppearance.Frameless. ([ni/nimble@ae9c35f](https://github.com/ni/nimble/commit/ae9c35fe75d93aa8b3a947759020b8558eb92490))
 
 ## 1.2.0